refactor: removed unused files, renamed validation files
