+++ package.json
@@ -35,7 +35,9 @@
     "reflect-metadata": "^0.1.13",
     "rimraf": "^3.0.2",
     "rxjs": "^7.2.0",
-    "uuid": "^8.3.2"
+    "uuid": "^8.3.2",
+    "@nestjs/swagger": "5.2.1",
+    "swagger-ui-express":"^5.0.0"
   },
   "devDependencies": {
     "@nestjs/cli": "^8.0.0",
