+++ src/api/artist/object-validation/validate-artist.ts
