+++ src/api/user/user.controller.ts
@@ -11,7 +11,7 @@ import { UUID } from 'crypto';
 import { checkUUID, notFound } from 'src/errorsAndMessages/errors';
 import { ICreateUserDto, IUpdatePasswordDto } from './interface/user.interface';
 import { UserService } from './user.service';
-import { isUserDataValid } from './dto/create-user.dto';
+import { isUserDataValid } from './object-validation/user-validation';
 
 @Controller()
 export class UserController {
